+++ apps/renderer/domains/editor/features/floating-toolbar/ui/FloatingToolbarContainer.vue
@@ -1,14 +1,14 @@
 <template>
-  <FloatingToolbar :show="isOpen" :position="position">
+  <FloatingToolbar :show="isOpen" :position="position" data-editor-floating-toolbar @mousedown.prevent>
     <template v-for="item in items" :key="item.id">
-      <div v-if="item.items" class="toolbar-group">
+      <ToolbarGroup v-if="item.items">
         <component
           v-for="button in item.items"
           :key="button.id"
           :is="button.component"
           v-bind="button.props"
         />
-      </div>
+      </ToolbarGroup>
       <component v-else :is="item.component" v-bind="item.props" />
     </template>
   </FloatingToolbar>
@@ -17,7 +17,7 @@
 <script setup>
 import { computed } from 'vue';
 import { floatingToolbarService } from '../service';
-import FloatingToolbar from './FloatingToolbar.vue';
+import { FloatingToolbar, ToolbarGroup } from '@linnya/renderer-ui';
 
 const isOpen = computed(() => floatingToolbarService.state.isOpen);
 const position = computed(() => floatingToolbarService.state.position);